Refactored isCacheableRequest and isCacheableResponse out of CacheFilter to a separate utility class for easier testing and extension
