+++ source/extensions/filters/http/cache/BUILD
@@ -18,12 +18,23 @@ envoy_cc_library(
     hdrs = ["cache_filter.h"],
     deps = [
         ":http_cache_lib",
+        ":cache_filter_utils_lib",
         "//source/common/common:logger_lib",
         "//source/common/common:macros",
         "//source/common/http:header_map_lib",
         "//source/common/http:headers_lib",
         "//source/extensions/filters/http/common:pass_through_filter_lib",
         "@envoy_api//envoy/extensions/filters/http/cache/v3alpha:pkg_cc_proto",
+    ],
+)
+
+envoy_cc_library(
+    name = "cache_filter_utils_lib",
+    srcs = ["cache_filter_utils.cc"],
+    hdrs = ["cache_filter_utils.h"],
+    deps = [
+        "//source/common/http:headers_lib",
+        "//source/common/common:utility_lib",
     ],
 )
