+++ source/extensions/filters/http/cache/cache_filter_utils.h
@@ -1,0 +1,21 @@
+#pragma once
+
+#include "common/http/headers.h"
+#include "common/common/utility.h"
+
+namespace Envoy {
+namespace Extensions {
+namespace HttpFilters {
+namespace Cache {
+class CacheFilterUtils {
+public:
+  // Checks if a request can be served from cache
+  static bool isCacheableRequest(Http::RequestHeaderMap& headers);
+
+  // Checks if a response can be stored in cache
+  static bool isCacheableResponse(Http::ResponseHeaderMap& headers);
+};
+} // namespace Cache
+} // namespace HttpFilters
+} // namespace Extensions
+} // namespace Envoy
